+++ packages/core/src/sva.ts
@@ -1,16 +1,6 @@
 import { cx } from "./class-utils";
-import type {
-  ClassProp,
-  CreateOptions,
-  MetaConfig,
-  RecordClassValue,
-  SlotVariants,
-  SVA,
-  SVAComponent,
-  SVAConfig,
-  SVAVariants
-} from "./types";
-import { falsyToString, hasProps, kebabCase } from "./utils";
+import type { ClassValue, CreateOptions } from "./types";
+import { type ClassProp, falsyToString, hasProps, kebabCase, type MetaConfig, type StringToBoolean } from "./utils";
 
 /* -----------------------------------------------------------------------------
  * SVA
@@ -245,3 +235,236 @@ export const createSVA = (options?: CreateOptions): SVA => {
  * ```
  */
 export const sva: SVA = createSVA();
+
+/* -----------------------------------------------------------------------------
+ * Types
+ * -----------------------------------------------------------------------------*/
+
+/**
+ * Reusable type for compound styles that apply based on multiple variant combinations.
+ *
+ * @template RCV - The type of record class values.
+ * @template V - The type of variants.
+ * @example
+ * ```ts
+ * type ButtonCompound = CompoundBase<
+ *   { root: string; item: string },
+ *   { size: { small: { root: string }; medium: { root: string } } }
+ * >;
+ * // ButtonCompound = { size?: "small" | "medium" | ("small" | "medium")[] };
+ * ```
+ */
+type CompoundBase<S extends string, RCV extends RecordClassValue<S>, V extends SlotVariants<S, RCV>> = {
+  [K in keyof V]?: StringToBoolean<keyof V[K]> | StringToBoolean<keyof V[K]>[];
+};
+
+/**
+ * Represents the default variants for a component.
+ *
+ * @template RCV - The type of record class values.
+ * @example
+ * ```ts
+ * {
+ *   size: {
+ *     small: {
+ *       root: "text-sm"
+ *     },
+ *     medium: {
+ *       root: "text-base"
+ *     }
+ *   }
+ * }
+ * ```
+ */
+type DefaultVariants<S extends string, RCV extends RecordClassValue<S>> = {
+  [key: string]: {
+    [key: string]: SlotsClassValue<S, RCV>;
+  };
+};
+
+/**
+ * Represents a dictionary where keys are strings and values are ClassValue.
+ */
+type RecordClassValue<S extends string> = Record<S, ClassValue>;
+
+/**
+ * Represents the class values for slots, where keys are slot names and values are class names.
+ *
+ * @template RCV - The type of record class values.
+ * @example
+ * ```ts
+ * type ButtonSlots = SlotsClassValue<{ root: string; item: string }>;
+ * // ButtonSlots = { root?: ClassValue; item?: ClassValue };
+ * ```
+ */
+type SlotsClassValue<S extends string, RCV extends RecordClassValue<S>> = {
+  [K in keyof RCV]?: ClassValue;
+};
+
+/**
+ * Represents the variants for a component.
+ *
+ * @template RCV - The type of record class values.
+ * @template V - The type of variants.
+ * @example
+ * ```ts
+ * type ButtonVariants = Variants<
+ *   {
+ *     root: string;
+ *     item: string;
+ *   },
+ *   {
+ *     size: {
+ *       small: {
+ *         root: string;
+ *       };
+ *       medium: {
+ *         root: string;
+ *       };
+ *     };
+ *   }
+ * >;
+ * // type ButtonVariants =
+ * // | { size?: { small?: { root?: ClassValue }; medium?: { root?: ClassValue } } }
+ * // | { size: { small: { root: ClassValue }; medium: { root: ClassValue } } };
+ * ```
+ */
+type SlotVariants<
+  S extends string,
+  RCV extends RecordClassValue<S>,
+  V extends DefaultVariants<S, RCV> = DefaultVariants<S, RCV>
+> =
+  | {
+      [K in keyof V]?: {
+        [K2 in keyof V[K]]?: SlotsClassValue<S, RCV>;
+      };
+    }
+  | DefaultVariants<S, RCV>;
+
+/**
+ * Represents the props for a component with variants and slots.
+ *
+ * @template RCV - The type of record class values.
+ * @template V - The type of variants.
+ * @example
+ * ```ts
+ * type ButtonProps = SVAVariants<
+ *   { root: string; item: string },
+ *   { size: { small: { root: string }; medium: { root: string } } }
+ * >;
+ * // ButtonProps = { size?: "small" | "medium" };
+ * ```
+ */
+type SVAVariants<S extends string, RCV extends RecordClassValue<S>, V extends SlotVariants<S, RCV>> = {
+  [K in keyof V]?: StringToBoolean<keyof V[K]>;
+};
+
+/**
+ * Represents the configuration object for the `sva` function.
+ *
+ * @template S - The name the of slots.
+ * @template RCV - The type of record class values.
+ * @property {string} [name] - Optional component name.
+ * @property {S=} [base] - Slots allow you to separate a component into multiple parts.
+ * @property {V=} variants - Variants allow you to create multiple versions of the same component.
+ * @property {Array<SVAVariants<S> & ( { class?: SlotsClassValue<RCV>; className?: never; } | { class?: never; className?: SlotsClassValue<RCV>; } )>=} compoundVariants - Compound variants allow you to apply classes to multiple variants at once.
+ * @property {Array<{ slots: Array<keyof S> } & SVAVariants<S, S> & ClassProp>=} compoundSlots - Compound slots allow you to apply classes to multiple slots at once.
+ * @property {SVAVariants<S>=} defaultVariants - Default variants allow you to set default variants for a component.
+ *
+ * @example
+ * ```ts
+ * const buttonConfig: SVAConfig<
+ *   { root: string; item: string },
+ *   { size: { small: { root: string }; medium: { root: string } } }
+ * > = {
+ *   slots: {
+ *     root: "flex",
+ *     item: "px-2 py-1"
+ *   },
+ *   variants: {
+ *     size: {
+ *       small: {
+ *         root: "text-sm"
+ *       },
+ *       medium: {
+ *         root: "text-base"
+ *       }
+ *     }
+ *   },
+ *   compoundVariants: [{ size: "small", class: { root: "font-bold" } }],
+ *   compoundSlots: [{ slots: ["root", "item"], class: "rounded" }],
+ *   defaultVariants: {
+ *     size: "medium"
+ *   }
+ * };
+ * ```
+ */
+type SVAConfig<S extends string, RCV extends Partial<RecordClassValue<S>>, V extends SlotVariants<S, RCV>> = {
+  /**
+   * The name of the sva component.
+   */
+  name?: string;
+  /**
+   * The base styles of the component.
+   */
+  base?: RCV;
+  /**
+   * The parts/slots of the component.
+   */
+  slots: S[] | Readonly<S[]>;
+  /**
+   * Variants allow you to create multiple versions of the same component.
+   */
+  variants?: V;
+  /**
+   * Compound variants allow you to apply classes to multiple variants at once.
+   */
+  compoundVariants?: Array<
+    CompoundBase<S, RCV, V> &
+      (
+        | {
+            class?: SlotsClassValue<S, RCV>;
+            className?: never;
+          }
+        | {
+            class?: never;
+            className?: SlotsClassValue<S, RCV>;
+          }
+      )
+  >;
+  /**
+   * Compound slots allow you to apply classes to multiple slots at once.
+   */
+  compoundSlots?: Array<{ slots: S[] | Readonly<S[]> } & CompoundBase<S, RCV, V> & ClassProp>;
+  /**
+   * Default variants allow you to set default variants for a component.
+   */
+  defaultVariants?: SVAVariants<S, RCV, V>;
+};
+
+/**
+ * Represents the return type of the SVA function.
+ *
+ * @template S - The name the of slots.
+ * @template RCV - The type of record class values.
+ * @returns {((props?: SVAVariants<S>) => { [K in keyof RCV]: (slotProps?: SVAVariants<S> & ClassProp) => string; }) & { metaConfig: MetaConfig; styleConfig: SVAConfig<S>; }} A function that generates slot-specific class names based on props.
+ */
+type SVAComponent<S extends string, RCV extends RecordClassValue<S>, SV extends SlotVariants<S, RCV>> = ((
+  props?: SVAVariants<S, RCV, SV>
+) => Record<S, (slotProps?: SVAVariants<S, RCV, SV> & ClassProp) => string>) & {
+  metaConfig: MetaConfig;
+  styleConfig: SVAConfig<S, RCV, SV>;
+};
+
+/**
+ * Creates a slots variants authority (SVA) function for a component.
+ *
+ * @template S - The name the of slots.
+ * @param {SVAConfig<S>} styleConfig - Configuration options for the SVA function.
+ * @param {MetaConfig} [metaConfig] - Add JSDoc and CSS comments to components when generating with `unplugin-jade-garden`.
+ * @returns {SVAComponent<S>} The return type of the SVA function.
+ */
+type SVA = <S extends string, RCV extends RecordClassValue<S>, SV extends SlotVariants<S, RCV>>(
+  styleConfig: SVAConfig<S, RCV, SV>,
+  metaConfig?: MetaConfig
+) => SVAComponent<S, RCV, SV>;
