+++ packages/core/src/types.ts
@@ -25,6 +25,11 @@ type Attribute<T> =
   : never;
 
 /**
+ * An array of {@link ClassValue}.
+ */
+type ClassArray = ClassValue[];
+
+/**
  * Represents the minimum structure to work with class names.
  * Fully compatible input for `jade-garden/class-utils` functions.
  */
@@ -36,7 +41,6 @@ export type ClassStrings = string | string[];
  * - It can be a `string`, `number`, `bigint`, `null`, `boolean`, `undefined`, an array of ClassValue, or a Record with values of `any`.
  */
 export type ClassValue = ClassArray | Record<string, any> | string | number | bigint | null | boolean | undefined;
-export type ClassArray = ClassValue[];
 
 /**
  * The options used to modify your custom merge functions (`createCVA` and `createSVA`).
@@ -109,45 +113,6 @@ export type CreateOptions = {
    * ```
    */
   useStylesheet?: boolean;
-};
-
-/**
- * **FOR LIBRARY AUTHORS**
- *
- * Add JSDoc and CSS comments to components when generating with `unplugin-jade-garden`.
- */
-export type MetaConfig = {
-  /**
-   * Adds a `deprecated` tag.
-   *
-   * Adds a description if `deprecated` is a string.
-   *
-   * @see https://jsdoc3.vercel.app/tags/deprecated
-   */
-  deprecated?: boolean | string;
-
-  /**
-   * Adds a `description` tag.
-   *
-   * @see https://jsdoc3.vercel.app/tags/description
-   */
-  description?: string;
-
-  /**
-   * Adds a `name` tag.
-   *
-   * `name` should be the same as `name` in style configuration.
-   *
-   * @see https://jsdoc3.vercel.app/tags/name
-   */
-  name?: string;
-
-  /**
-   * Adds a `see` tag.
-   *
-   * @see https://jsdoc3.vercel.app/tags/see
-   */
-  see?: string;
 };
 
 /**
@@ -224,405 +189,3 @@ export type VariantProps<Component extends (...args: any) => any> = Omit<
   OmitUndefined<Parameters<Component>[0]>,
   "class" | "className"
 >;
-
-/* -----------------------------------------------------------------------------
- * BELOW IS INTENDED FOR INTERNAL USE ONLY
- * -----------------------------------------------------------------------------*/
-
-/* -----------------------------------------------------------------------------
- * Types
- * -----------------------------------------------------------------------------*/
-
-/**
- * Represents the `class` and `className` props for `cva` and `sva`.
- * Ensures that only one of `class` or `className` is present.
- */
-export type ClassProp =
-  | {
-      class?: ClassValue;
-      className?: never;
-    }
-  | {
-      class?: never;
-      className?: ClassValue;
-    };
-
-/**
- * Represents a dictionary where keys are strings and values are ClassValue.
- */
-export type RecordClassValue<S extends string> = Record<S, ClassValue>;
-
-/**
- * Converts "true" or "false" string literals to boolean types.
- * Otherwise, returns the original type.
- *
- * @template T - The type to convert.
- * @returns {T extends "true" | "false" ? boolean : T} The converted type.
- */
-export type StringToBoolean<T> = T extends "true" | "false" ? boolean : T;
-
-/* -----------------------------------------------------------------------------
- * CVA
- * -----------------------------------------------------------------------------*/
-
-/**
- * Represents the variant configurations for `cva`.
- * Each variant is a record of class names, keyed by variant values.
- *
- * @example
- * ```ts
- * {
- *   size: {
- *     small: "text-sm py-1 px-2",
- *     medium: "text-base py-2 px-4",
- *   },
- *   intent: {
- *     primary: "bg-blue-500 text-white",
- *     secondary: "bg-gray-200 text-gray-800",
- *   }
- * }
- * ```
- */
-export type Variant = Record<string, RecordClassValue<string>>;
-
-/**
- * Represents the schema for variant props in `cva`.
- * Each variant prop is typed as a StringToBoolean of the corresponding variant value keys.
- *
- * @template V - The type of variants.
- * @example
- * ```ts
- * type ButtonVariants = CVAVariants<{
- *   size: { small: string; medium: string };
- *   intent: { primary: string; secondary: string };
- * }>;
- * // ButtonVariants = { size?: "small" | "medium"; intent?: "primary" | "secondary" };
- * ```
- */
-export type CVAVariants<V extends Variant> = {
-  [K in keyof V]?: StringToBoolean<keyof V[K]>;
-};
-
-/**
- * Represents the configuration object for the `cva` function.
- *
- * @template V - The type of variants.
- * @property {string=} name - Optional component name.
- * @property {ClassValue=} base - The base class name for the component.
- * @property {V=} variants - Variants allow you to create multiple versions of the same component.
- * @property {(V extends Variant ? (CVAVariants<V> | { [K in keyof V]?: StringToBoolean<keyof V[K]> | StringToBoolean<keyof V[K]>[]; }) & ClassProp : ClassProp)[]=} compoundVariants - Compound variants allow you to apply classes to multiple variants at once.
- * @property {CVAVariants<V>=} defaultVariants - Default variants allow you to set default variants for a component.
- *
- * @example
- * ```ts
- * const buttonConfig: CVAConfig<{
- *   size: { small: string; medium: string };
- *   intent: { primary: string; secondary: string };
- * }> = {
- *   base: "rounded-md",
- *   variants: {
- *     size: {
- *       small: "text-sm",
- *       medium: "text-base"
- *     },
- *     intent: {
- *       primary: "bg-blue-500",
- *       secondary: "bg-gray-200"
- *     }
- *   },
- *   compoundVariants: [
- *     {
- *       size: "small",
- *       intent: "primary",
- *       class: "font-bold"
- *     }
- *   ],
- *   defaultVariants: {
- *     size: "medium",
- *     intent: "primary"
- *   }
- * };
- * ```
- */
-export type CVAConfig<V extends Variant> = {
-  /**
-   * The name of the cva component.
-   */
-  name?: string;
-  /**
-   * The base class name for the component.
-   */
-  base?: ClassValue;
-  /**
-   * Variants allow you to create multiple versions of the same component.
-   */
-  variants?: V;
-  /**
-   * Compound variants allow you to apply classes to multiple variants at once.
-   */
-  compoundVariants?: (V extends Variant
-    ? (
-        | CVAVariants<V>
-        | {
-            [K in keyof V]?: StringToBoolean<keyof V[K]> | StringToBoolean<keyof V[K]>[];
-          }
-      ) &
-        ClassProp
-    : ClassProp)[];
-  /**
-   * Default variants allow you to set default variants for a component.
-   */
-  defaultVariants?: CVAVariants<V>;
-};
-
-/**
- * Represents the return type of the CVA function.
- *
- * @template V - The type of variants.
- * @returns {((props?: V extends Variant ? CVAVariants<V> & ClassProp : ClassProp) => string)& { metaConfig: MetaConfig; styleConfig: CVAConfig<V>; }} A function that generates class names based on props.
- */
-export type CVAComponent<V extends Variant> = ((
-  props?: V extends Variant ? CVAVariants<V> & ClassProp : ClassProp
-) => string) & {
-  metaConfig: MetaConfig;
-  styleConfig: CVAConfig<V>;
-};
-
-/**
- * Creates a class variant authority (cva) function.
- * Generates class names based on provided variants and props.
- *
- * @template V - The type of variants.
- * @param {CVAConfig<V>} styleConfig - The cva configuration object.
- * @param {MetaConfig} [metaConfig] - Add JSDoc and CSS comments to components when generating with `unplugin-jade-garden`.
- * @returns {CVAComponent<V>} A function that generates class names based on props.
- */
-export type CVA = <V extends Variant = {}>(styleConfig: CVAConfig<V>, metaConfig?: MetaConfig) => CVAComponent<V>;
-
-/* -----------------------------------------------------------------------------
- * SVA
- * -----------------------------------------------------------------------------*/
-
-/**
- * Reusable type for compound styles that apply based on multiple variant combinations.
- *
- * @template RCV - The type of record class values.
- * @template V - The type of variants.
- * @example
- * ```ts
- * type ButtonCompound = CompoundBase<
- *   { root: string; item: string },
- *   { size: { small: { root: string }; medium: { root: string } } }
- * >;
- * // ButtonCompound = { size?: "small" | "medium" | ("small" | "medium")[] };
- * ```
- */
-type CompoundBase<S extends string, RCV extends RecordClassValue<S>, V extends SlotVariants<S, RCV>> = {
-  [K in keyof V]?: StringToBoolean<keyof V[K]> | StringToBoolean<keyof V[K]>[];
-};
-
-/**
- * Represents the default variants for a component.
- *
- * @template RCV - The type of record class values.
- * @example
- * ```ts
- * {
- *   size: {
- *     small: {
- *       root: "text-sm"
- *     },
- *     medium: {
- *       root: "text-base"
- *     }
- *   }
- * }
- * ```
- */
-type DefaultVariants<S extends string, RCV extends RecordClassValue<S>> = {
-  [key: string]: {
-    [key: string]: SlotsClassValue<S, RCV>;
-  };
-};
-
-/**
- * Represents the class values for slots, where keys are slot names and values are class names.
- *
- * @template RCV - The type of record class values.
- * @example
- * ```ts
- * type ButtonSlots = SlotsClassValue<{ root: string; item: string }>;
- * // ButtonSlots = { root?: ClassValue; item?: ClassValue };
- * ```
- */
-type SlotsClassValue<S extends string, RCV extends RecordClassValue<S>> = {
-  [K in keyof RCV]?: ClassValue;
-};
-
-/**
- * Represents the variants for a component.
- *
- * @template RCV - The type of record class values.
- * @template V - The type of variants.
- * @example
- * ```ts
- * type ButtonVariants = Variants<
- *   {
- *     root: string;
- *     item: string;
- *   },
- *   {
- *     size: {
- *       small: {
- *         root: string;
- *       };
- *       medium: {
- *         root: string;
- *       };
- *     };
- *   }
- * >;
- * // type ButtonVariants =
- * // | { size?: { small?: { root?: ClassValue }; medium?: { root?: ClassValue } } }
- * // | { size: { small: { root: ClassValue }; medium: { root: ClassValue } } };
- * ```
- */
-export type SlotVariants<
-  S extends string,
-  RCV extends RecordClassValue<S>,
-  V extends DefaultVariants<S, RCV> = DefaultVariants<S, RCV>
-> =
-  | {
-      [K in keyof V]?: {
-        [K2 in keyof V[K]]?: SlotsClassValue<S, RCV>;
-      };
-    }
-  | DefaultVariants<S, RCV>;
-
-/**
- * Represents the props for a component with variants and slots.
- *
- * @template RCV - The type of record class values.
- * @template V - The type of variants.
- * @example
- * ```ts
- * type ButtonProps = SVAVariants<
- *   { root: string; item: string },
- *   { size: { small: { root: string }; medium: { root: string } } }
- * >;
- * // ButtonProps = { size?: "small" | "medium" };
- * ```
- */
-export type SVAVariants<S extends string, RCV extends RecordClassValue<S>, V extends SlotVariants<S, RCV>> = {
-  [K in keyof V]?: StringToBoolean<keyof V[K]>;
-};
-
-/**
- * Represents the configuration object for the `sva` function.
- *
- * @template S - The name the of slots.
- * @template RCV - The type of record class values.
- * @property {string} [name] - Optional component name.
- * @property {S=} [base] - Slots allow you to separate a component into multiple parts.
- * @property {V=} variants - Variants allow you to create multiple versions of the same component.
- * @property {Array<SVAVariants<S> & ( { class?: SlotsClassValue<RCV>; className?: never; } | { class?: never; className?: SlotsClassValue<RCV>; } )>=} compoundVariants - Compound variants allow you to apply classes to multiple variants at once.
- * @property {Array<{ slots: Array<keyof S> } & SVAVariants<S, S> & ClassProp>=} compoundSlots - Compound slots allow you to apply classes to multiple slots at once.
- * @property {SVAVariants<S>=} defaultVariants - Default variants allow you to set default variants for a component.
- *
- * @example
- * ```ts
- * const buttonConfig: SVAConfig<
- *   { root: string; item: string },
- *   { size: { small: { root: string }; medium: { root: string } } }
- * > = {
- *   slots: {
- *     root: "flex",
- *     item: "px-2 py-1"
- *   },
- *   variants: {
- *     size: {
- *       small: {
- *         root: "text-sm"
- *       },
- *       medium: {
- *         root: "text-base"
- *       }
- *     }
- *   },
- *   compoundVariants: [{ size: "small", class: { root: "font-bold" } }],
- *   compoundSlots: [{ slots: ["root", "item"], class: "rounded" }],
- *   defaultVariants: {
- *     size: "medium"
- *   }
- * };
- * ```
- */
-export type SVAConfig<S extends string, RCV extends Partial<RecordClassValue<S>>, V extends SlotVariants<S, RCV>> = {
-  /**
-   * The name of the sva component.
-   */
-  name?: string;
-  /**
-   * The base styles of the component.
-   */
-  base?: RCV;
-  /**
-   * The parts/slots of the component.
-   */
-  slots: S[] | Readonly<S[]>;
-  /**
-   * Variants allow you to create multiple versions of the same component.
-   */
-  variants?: V;
-  /**
-   * Compound variants allow you to apply classes to multiple variants at once.
-   */
-  compoundVariants?: Array<
-    CompoundBase<S, RCV, V> &
-      (
-        | {
-            class?: SlotsClassValue<S, RCV>;
-            className?: never;
-          }
-        | {
-            class?: never;
-            className?: SlotsClassValue<S, RCV>;
-          }
-      )
-  >;
-  /**
-   * Compound slots allow you to apply classes to multiple slots at once.
-   */
-  compoundSlots?: Array<{ slots: S[] | Readonly<S[]> } & CompoundBase<S, RCV, V> & ClassProp>;
-  /**
-   * Default variants allow you to set default variants for a component.
-   */
-  defaultVariants?: SVAVariants<S, RCV, V>;
-};
-
-/**
- * Represents the return type of the SVA function.
- *
- * @template S - The name the of slots.
- * @template RCV - The type of record class values.
- * @returns {((props?: SVAVariants<S>) => { [K in keyof RCV]: (slotProps?: SVAVariants<S> & ClassProp) => string; }) & { metaConfig: MetaConfig; styleConfig: SVAConfig<S>; }} A function that generates slot-specific class names based on props.
- */
-export type SVAComponent<S extends string, RCV extends RecordClassValue<S>, SV extends SlotVariants<S, RCV>> = ((
-  props?: SVAVariants<S, RCV, SV>
-) => Record<S, (slotProps?: SVAVariants<S, RCV, SV> & ClassProp) => string>) & {
-  metaConfig: MetaConfig;
-  styleConfig: SVAConfig<S, RCV, SV>;
-};
-
-/**
- * Creates a slots variants authority (SVA) function for a component.
- *
- * @template S - The name the of slots.
- * @param {SVAConfig<S>} styleConfig - Configuration options for the SVA function.
- * @param {MetaConfig} [metaConfig] - Add JSDoc and CSS comments to components when generating with `unplugin-jade-garden`.
- * @returns {SVAComponent<S>} The return type of the SVA function.
- */
-export type SVA = <S extends string, RCV extends RecordClassValue<S>, SV extends SlotVariants<S, RCV>>(
-  styleConfig: SVAConfig<S, RCV, SV>,
-  metaConfig?: MetaConfig
-) => SVAComponent<S, RCV, SV>;
